requirements.txt
Django>=3.0,<4.0
APScheduler>=3.6
djangorestframework>=3.1